requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.25.3
Pygments==2.18.0
PyMatting==1.1.12
pyreadline3==3.4.1
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
sympy==1.13.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.7.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
